Core: move polite sync helpers to utils
